reorg: pull bslv into separate source files
